Trying out some unique limits reasoning
I recognize that the gap between univalence and unique limits is less
clear to me. I want to punt on that for now and complete the rest of
the proof. I'm tightening up the generality of the universe levels,
I'm not sure yet if that's right or not. I suspect it's not.